+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Unreleased
   around
 - Introduced `normalize::Reason` enum to provide best guess at why normalization
   was not successful as part of the `normalize::UserMeta::Unknown` variant
+- Added `From<Box<dyn std::error::Error>>` conversion to `Error`
 - Reduced number of allocations performed on address normalization and
   process symbolization paths
 - Improved symbolization performance on certain unoptimized DWARF data